requirements.txt
amqp==5.0.6
asgiref==3.3.4
backcall==0.2.0
billiard==3.6.4.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
decorator==5.0.9
Django==3.2.4
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
json5==0.9.5
kombu==5.1.0
matplotlib-inline==0.1.2
mysqlclient @ file:///C:/Users/SPURGE/ready_assist_file_upload/file_upload/mysqlclient-1.4.6-cp39-cp39-win_amd64.whl
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.4
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
Pygments==2.9.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
traitlets==5.0.5
vine==5.0.0
wcwidth==0.2.5
